TCS Specific Trainning Questions
11.A company wishes to encode its data is in the form of a string and is case sensitive. They wish to encode the data with respect to a specific character. They wish to count the number of times the character reoccurs in the given data so that they can encode the data accordingly.

Write an algorithm to find the count of the specific character in the given data.
Input
The first line of the input consists of a string data representing the data to be encoded.
The next line of the input consists of a character coder representing the character to be counted in the data.
Output
Print an integer representing the count of the specific character.


Example Input:
haveagoodday a
Output:
3
Explanation:
The character “a” occurs thrice in the data. So, the output is 3.